@@ -52,6 +52,6 @@ class VerifyServiceProvider extends ServiceProvider
      */
     public function provides()
     {
-        return array();
+        return [];
     }
 }
